+++ packages/config/src/index.ts
@@ -1,4 +1,4 @@
 export * from './bridges'
 export * from './common'
 export * from './layer2s'
-export * from './tokens'
+export * from './tokens/tokens'
